it-tickets-kaggle-lighning-mlflow-test/requirements.txt
lightning
pyarrow
pandas
scikit-learn
transformers
torchmetrics
mlflow
loguru
google-cloud-storage
pyspark
delta